requirements.txt
Django
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
pytest-django
pdbpp
fancycompleter
